+++ src/main/java/br/com/jorgemelo/nimbusfilemanager/metadata/application/MetadataRebuildAsyncRunner.java
@@ -84,7 +84,7 @@ public class MetadataRebuildAsyncRunner {
 	}
 
 	/** 0-100, or -1 when the total is unknown. */
-	public int percent() {
+	public double percent() {
 		return ProgressMath.percent(processed.get(), total.get());
 	}
